requirements.txt
Flask>=2.0.0
flask-cors
yt-dlp
opencv-python-headless
numpy
fpdf2
Pillow